Updated GOMstreamer to work with the new GSL format. Among the changes are the following:
Logins are done via a different URL and with slightly
different parameters.

Different stream qualities are now passed in as parameters when
requesting the GOX XML. The strLevel parameter takes 'HQ', 'SQ' and
'SQTest'. The first two require a premium account.

The HTML now needs to be parsed to find the title of the stream too,
however this is probably not necessary. Regardless, the parsing is now
undertaken.

The GOX XML file that is requested and parsed is much the same as
before, no real changes required.